refactor: return type
