+++ backend/src/main/java/com/SeeAndYouGo/SeeAndYouGo/OAuth/dto/TokenDto.java
@@ -1,0 +1,10 @@
+package com.SeeAndYouGo.SeeAndYouGo.OAuth.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class TokenDto {
+    private final String token;
+}
